Move code to access ttpwb API to a separate file
